+++ CHANGELOG.md
@@ -16,6 +16,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Fixed
 
 - Allow menu items to be triggered by repeating keyboard events.
+- Prevent cursor flickering when resizing the palette.
 
 ## [1.0.0] - 2022-08-08